# connectease_backend/openapi.yaml
openapi: 3.0.3
info:
  title: ConnectEase API
  description: >
    REST API for ConnectEase — a service marketplace connecting customers with
    vendors for home services, PGs, food, electrical, plumbing, cleaning, and laundry.

    **Authentication:** All protected endpoints read a JWT from an HttpOnly cookie
    named `jwt`. Call `POST /api/auth/signin` first — the cookie is set automatically.

    **Roles:**
    - `customer` — can browse, review, chat, post to community
    - `vendor`   — same as customer plus full service listing management (`/api/vendor/**`)
  version: 1.0.0
  contact:
    name: ConnectEase Team
servers:
  - url: http://localhost:8081
    description: Local development server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: >
        HttpOnly JWT cookie set by POST /api/auth/signin.
        In Swagger UI click "Authorize", paste the raw JWT value.
        In Postman the cookie is stored automatically after login.

# ─────────────────────────────────────────────────────────────────────────────
# SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: Operation completed successfully
...
    ImageDTO:
      type: object
      properties:
        url:
          type: string
          example: https://cdn.example.com/services/cleaning1.jpg
        isPrimary:
          type: boolean
          example: true
    ReviewDTO:
      type: object
      properties:
        userName:
          type: string
          example: Ravi Kumar
        profileImage:
          type: string
          nullable: true
          example: https://cdn.example.com/avatars/ravi.jpg
        review:
          type: string
          example: Excellent service, very thorough!
        score:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
    ListingCardDTO:
      type: object
      properties:
        sid:
          type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Home Deep Cleaning
        description:
          type: string
          example: Professional full-home cleaning service
        price:
          type: number
          format: double
          example: 1500.00
        categoryName:
          type: string
          example: cleaning service
        city:
          type: string
          example: Hyderabad
        area:
          type: string
          example: Gachibowli
        primaryImageUrl:
          type: string
          nullable: true
          example: https://cdn.example.com/services/cleaning.jpg
    ServiceDetailsDTO:
      type: object
      properties:
        sid:
          type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
        name:
          type: string
          example: Home Deep Cleaning
        description:
          type: string
          example: Full home deep cleaning including kitchen and bathrooms
        price:
          type: number
          format: double
          example: 1500.00
        totalViews:
          type: integer
          format: int64
          example: 243
        vendorName:
          type: string
          example: CleanPro Services
        averageRating:
          type: number
          format: double
          example: 4.3
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ReviewDTO'
        images:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ImageDTO'
...
    CategoryRef:
      type: object
      required:
        - cid
      properties:
        cid:
          type: string
          description: UUID of an existing category
          example: cat-uuid-from-database
    LocationInput:
      type: object
      properties:
        address:
          type: string
          example: "123 Main Street"
        city:
          type: string
          example: Hyderabad
        area:
          type: string
          example: Gachibowli
        latitude:
          type: number
          format: double
          example: 17.44086700
        longitude:
          type: number
          format: double
          example: 78.34875100
    FeatureInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Feature is created if it does not already exist (case-insensitive match)
          example: Eco-friendly products
    ServiceImageInput:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          example: https://cdn.example.com/services/cleaning1.jpg
        isPrimary:
          type: boolean
          example: true
    ServiceCreateRequest:
      type: object
      required:
        - name
        - price
      properties:
        name:
          type: string
          example: Home Deep Cleaning
        description:
          type: string
          example: Full home deep cleaning including kitchen and bathrooms
        price:
          type: number
          format: double
          example: 1500.00
        category:
          $ref: '#/components/schemas/CategoryRef'
        location:
          $ref: '#/components/schemas/LocationInput'
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureInput'
        images:
          type: array
          items:
            $ref: '#/components/schemas/ServiceImageInput'
    ServiceUpdateRequest:
      type: object
      description: All fields are optional. Only provided fields are updated.
      properties:
        name:
          type: string
          example: Premium Home Deep Cleaning
        description:
          type: string
          example: Updated description with more detail
        price:
          type: number
          format: double
          example: 1800.00
        category:
          $ref: '#/components/schemas/CategoryRef'
        features:
          type: array
          items:
            $ref: '#/components/schemas/FeatureInput'
    VendorDashboardDTO:
      type: object
      properties:
        vendorName:
          type: string
          example: CleanPro Services
        activeListings:
          type: integer
          format: int64
          example: 5
        totalViews:
          type: integer
          format: int64
          example: 1240
        totalReviews:
          type: integer
          format: int64
          example: 28
        averageRating:
          type: number
          format: double
          example: 4.2

    # ── Community ─────────────────────────────────────────────────────────────
...
    PageListingCardDTO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ListingCardDTO'
        totalElements:
          type: integer
          example: 25
        totalPages:
          type: integer
          example: 3
        size:
          type: integer
          example: 10
        number:
          type: integer
          description: Current page index (0-based)
          example: 0
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
...
paths:
...
  /api/services:
    get:
      tags: [Services]
      summary: Browse all services (paginated, sorted newest first)
      operationId: getAllServices
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Paginated service cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageListingCardDTO'
              example:
                content:
                  - sid: s1b2c3d4-e5f6-7890-abcd-ef1234567890
                    name: Home Deep Cleaning
                    description: Professional full-home cleaning
                    price: 1500.00
                    categoryName: cleaning service
                    city: Hyderabad
                    area: Gachibowli
                    primaryImageUrl: https://cdn.example.com/services/cleaning.jpg
                totalElements: 25
                totalPages: 3
                size: 10
                number: 0
                first: true
                last: false
...
  /api/vendor/dashboard:
    get:
      tags: [Vendor]
      summary: Get vendor dashboard statistics
      operationId: getVendorDashboard
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Aggregated stats for the logged-in vendor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorDashboardDTO'
              example:
                vendorName: CleanPro Services
                activeListings: 5
                totalViews: 1240
                totalReviews: 28
                averageRating: 4.2
        '403':
          description: Forbidden — must be logged in as a vendor
  /api/vendor/service/add:
    post:
      tags: [Vendor]
      summary: Create a new service listing
      operationId: createService
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceCreateRequest'
            example:
              name: Home Deep Cleaning
              description: Full home deep cleaning including kitchen and bathrooms
              price: 1500.00
              category:
                cid: cat-uuid-from-database
              location:
                address: "123 Main Street"
                city: Hyderabad
                area: Gachibowli
                latitude: 17.44086700
                longitude: 78.34875100
              features:
                - name: Eco-friendly products
                - name: 2-hour service
              images:
                - url: https://cdn.example.com/services/cleaning1.jpg
                  isPrimary: true
                - url: https://cdn.example.com/services/cleaning2.jpg
                  isPrimary: false
      responses:
        '200':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailsDTO'
              example:
                sid: s1b2c3d4-e5f6-7890-abcd-ef1234567890
                name: Home Deep Cleaning
                description: Full home deep cleaning including kitchen and bathrooms
                price: 1500.00
                totalViews: 0
                vendorName: CleanPro Services
                averageRating: 0.0
                reviews: []
                images: null
        '403':
          description: Forbidden — must be logged in as a vendor
        '500':
          description: Category or vendor not found
  /api/vendor/services:
    get:
      tags: [Vendor]
      summary: Get the logged-in vendor's own service listings
      operationId: getMyServices
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Vendor's own listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListingCardDTO'
              example:
                - sid: s1b2c3d4-e5f6-7890-abcd-ef1234567890
                  name: Home Deep Cleaning
                  description: Professional full-home cleaning
                  price: 1500.00
                  categoryName: cleaning service
                  city: Hyderabad
                  area: Gachibowli
                  primaryImageUrl: https://cdn.example.com/services/cleaning1.jpg
        '403':
          description: Forbidden — must be logged in as a vendor
  /api/vendor/service/{sid}:
    put:
      tags: [Vendor]
      summary: Update a service listing
      operationId: updateService
      description: All fields are optional. Only provided fields are updated.
      security:
        - cookieAuth: []
      parameters:
        - name: sid
          in: path
          required: true
          schema:
            type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
            example:
              name: Premium Home Deep Cleaning
              description: Updated description with more detail
              price: 1800.00
              category:
                cid: cat-uuid-from-database
              features:
                - name: Eco-friendly products
                - name: 3-hour service
                - name: Free follow-up visit
      responses:
        '200':
          description: Updated service details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailsDTO'
        '403':
          description: Forbidden — must be logged in as a vendor
        '500':
          description: Service not found or trying to update another vendor's service

    delete:
      tags: [Vendor]
      summary: Delete a service listing
      operationId: deleteService
      security:
        - cookieAuth: []
      parameters:
        - name: sid
          in: path
          required: true
          schema:
            type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '200':
          description: Service deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                status: success
                message: Service deleted successfully
        '403':
          description: Forbidden — must be logged in as a vendor
        '500':
          description: Service not found or trying to delete another vendor's service
  /api/vendor/service/{sid}/status:
    patch:
      tags: [Vendor]
      summary: Toggle service active/inactive status
      operationId: toggleServiceStatus
      description: >
        Each call flips the current `active` flag.
        First call → inactive. Second call → active again.
        No request body needed.
      security:
        - cookieAuth: []
      parameters:
        - name: sid
          in: path
          required: true
          schema:
            type: string
          example: s1b2c3d4-e5f6-7890-abcd-ef1234567890
      responses:
        '200':
          description: Service with updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailsDTO'
        '403':
          description: Forbidden — must be logged in as a vendor
        '500':
          description: Service not found or unauthorized

  # ════════════════════════════════════════════════════════════════════════════
  # RATINGS
  # ════════════════════════════════════════════════════════════════════════════
...